requirements.txt
cycler==0.10.0
future==0.18.2
iso8601==0.1.12
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.4
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.14.0
